TICOBackend/requirements.txt
celery==5.2.3
djangorestframework==3.13.1
django-celery-beat==2.2.1
requests==2.27.1